extract the RouteConfigFactory to a seperate class
